Modularization Mostly Complete
Almost all compatibility checks have now been moved to there own class called CompatCheck, the funcitons not moved are the Architecture Checkers, identifier checker has been moved the to main() to function, Prossesor Architecture checker still needs to be moved. I dont understand enough about the function to do this. Once that it done we are almost ready for submition